Add PIDPunzi
PIDPunzi computes the figure of merit out of a string of text files
(generated either by ComputeValue or PIDTable) separated by blank
spaces. It assumes a square table and hard-codely detects if Magnet
discrimnation is being done. It takes then the initial yields of each
channel (except the signal, which is not needed) and computs the FOM
for each combination.