PIDTable can now deal with weights
Since this script is designed to deal with the PIDCalib output ntuples,
the weights have to be taken from a second ntuple (essentially, the one
which PIDCalib was run over).

So the script, if no weights are requested, runs as usual. If weights
are requested, then a second file is read and the efficiency is computed
through an event loop (can't do better than this for now)