Added two functions related to fit yields
WriteFromWS.C: It takes a workspace, extracts some variables and writes
it in a file. Simple but this way I can get back to the yield file
structure.

ScaleYields.C: It rescales a set of yields so that the sum coincides
with the new total. This is used to rescale the yields of KstG and
LambdaG once the fit to each set is done.