Fix ROOT missing Links. Episode 1
The compiled version of ROOT provided by lxplus does not contain all the
libraries needed. There are some missing (Davix, btt, dvt...).

The log from when ROOT was built points out to certain versions of these
libraries, which have been manually added.

This is episode 1, as we are using ROOT 6.18. The idea is to move on to
ROOT 20, which is episode 2.

WHO FUCKED THIS UP!?!?!